Commit to the issue. [#70]
Russia diverted 34 trucks from an aid convoy into eastern Ukraine after
Russian and Ukrainian customs officials cleared them under the
assumption that the International Committee of the Red Cross would be
with them, the Ukrainian military said.
But the Red Cross said it was no longer with the convoy because of the
"volatile security situation," a reference to fighting between
pro-Russian rebels and Ukrainian forces.

This was a violation of the deal between the two countries that any
humanitarian convoy must be monitored by the Red Cross.